correct - apply delete_user to models.py

+++ profiles/models.py
@@ -1,7 +1,7 @@
 from django.db import models
-from django.db.models.signals import post_save
+from django.db.models.signals import post_save, post_delete
 from django.contrib.auth.models import User
-from django.db.models.signals import post_delete
+from django.dispatch import receiver
 
 # Profile Model
 class Profile(models.Model):
@@ -27,8 +27,9 @@ def create_profile(sender, instance, created, **kwargs):
 
 post_save.connect(create_profile, sender=User)
 
-
-def delete_user(sender, instance, using, **kwargs):
-    instance.owner.delete()
+@receiver(post_delete, sender=Profile)
+def delete_user(sender, instance, **kwargs):
+    if instance.owner:
+        instance.owner.delete()
 
 post_delete.connect(delete_user, sender=Profile)
